+++ helper/nr-epc-helper.cc
@@ -9,6 +9,7 @@
 
 #include "nr-epc-helper.h"
 
+#include "ns3/abort.h"
 #include "ns3/ipv4-address.h"
 #include "ns3/ipv6-address.h"
 #include "ns3/log.h"
@@ -43,6 +44,23 @@ NrEpcHelper::DoDispose()
 {
     NS_LOG_FUNCTION(this);
     Object::DoDispose();
+}
+
+uint8_t
+NrEpcHelper::ActivateUnstructuredQosFlow(Ptr<NetDevice> ueNrDevice,
+                                         uint64_t imsi,
+                                         uint16_t protocolNumber,
+                                         NrQosFlow flow)
+{
+    NS_LOG_FUNCTION(this << ueNrDevice << imsi << protocolNumber);
+    NS_ABORT_MSG("Unstructured PDU sessions are not supported by this EPC helper");
+}
+
+Ptr<VirtualNetDevice>
+NrEpcHelper::GetUnstructuredSessionDevice(uint64_t imsi, uint8_t qfi) const
+{
+    NS_LOG_FUNCTION(this << imsi << qfi);
+    NS_ABORT_MSG("Unstructured PDU sessions are not supported by this EPC helper");
 }
 
 } // namespace ns3