+++ model/nr-net-device.h
@@ -81,13 +81,31 @@ class NR_EXPORT NrNetDevice : public NetDevice
     TracedCallback<Ptr<const Packet>> m_rxTrace;   ///< Traced Callback for received packets
     TracedCallback<Ptr<const Packet>> m_dropTrace; ///< Traced Callback for dropped packets
     NetDevice::ReceiveCallback m_rxCallback;
+    Ptr<ErrorModel> m_receiveErrorModel; ///< Error model for receive packet events
 
     virtual bool DoSend(Ptr<Packet> packet, const Address& dest, uint16_t protocolNumber) = 0;
+
+    /**
+     * @brief Handle a received packet that is neither IPv4 nor IPv6.
+     *
+     * The default drops and traces the packet. A device that carries protocols
+     * other than IP overrides this to hand the packet up.
+     *
+     * @param p the packet
+     */
+    virtual void ForwardUnclassifiedUp(Ptr<Packet> p);
+
+    /**
+     * @brief Hand a received packet up to the node.
+     *
+     * @param p the packet
+     * @param protocolNumber the network layer protocol of the packet
+     */
+    void ForwardUp(Ptr<Packet> p, uint16_t protocolNumber);
 
   private:
     Mac48Address m_macAddress;
     Ptr<Node> m_node;
-    Ptr<ErrorModel> m_receiveErrorModel; ///< Error model for receive packet events
     mutable uint16_t m_mtu;
     bool m_linkUp;
     uint32_t m_ifIndex;